# docs/swagger-doc-export.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/airports": {
      "get": {
        "tags": [
          "airport-controller"
        ],
        "operationId": "getAirports",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AirportGetResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "airport-controller"
        ],
        "operationId": "updateAirport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AirportUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AirportUpdateResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "airport-controller"
        ],
        "operationId": "addAirport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AirportSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AirportSaveResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/airplanes": {
      "get": {
        "tags": [
          "airplane-controller"
        ],
        "operationId": "getAirplanes_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AirplaneGetResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "airplane-controller"
        ],
        "operationId": "updateAirplane",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AirplaneUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AirplaneUpdateResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "airplane-controller"
        ],
        "operationId": "addAirline_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AirplaneSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AirplaneSaveResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/airline-companies": {
      "get": {
        "tags": [
          "airline-company-controller"
        ],
        "operationId": "getAirlineCompanies",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AirlineCompanyGetResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "airline-company-controller"
        ],
        "operationId": "updateAirlineCompany",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AirlineCompanyUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AirlineCompanyUpdateResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "airline-company-controller"
        ],
        "operationId": "addAirlineCompany",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AirlineCompanySaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AirlineCompanySaveResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets": {
      "get": {
        "tags": [
          "ticket-controller"
        ],
        "operationId": "getTickets",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketGetResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ticket-controller"
        ],
        "operationId": "addTicket",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TicketSaveResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/flights": {
      "get": {
        "tags": [
          "flight-controller"
        ],
        "operationId": "getAirplanes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlightGetResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "flight-controller"
        ],
        "operationId": "addAirline",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlightSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FlightSaveResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/search": {
      "get": {
        "tags": [
          "ticket-controller"
        ],
        "operationId": "searchTicket",
        "parameters": [
          {
            "name": "ticketNumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TicketGetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/flights/search": {
      "get": {
        "tags": [
          "flight-controller"
        ],
        "operationId": "searchFlight",
        "parameters": [
          {
            "name": "origin",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "destination",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlightGetResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/airports/{id}": {
      "get": {
        "tags": [
          "airport-controller"
        ],
        "operationId": "getAirportById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AirportGetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/airports/search": {
      "get": {
        "tags": [
          "airport-controller"
        ],
        "operationId": "searchAirport",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AirportGetResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/airplanes/{id}": {
      "get": {
        "tags": [
          "airplane-controller"
        ],
        "operationId": "getAirplaneById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AirplaneGetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/airplanes/search": {
      "get": {
        "tags": [
          "airplane-controller"
        ],
        "operationId": "searchAirplane",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AirplaneGetResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/airline-companies/{id}": {
      "get": {
        "tags": [
          "airline-company-controller"
        ],
        "operationId": "getAirlineCompanyById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AirlineCompanyGetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/airline-companies/search": {
      "get": {
        "tags": [
          "airline-company-controller"
        ],
        "operationId": "searchAirline",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AirlineCompanyGetResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{ticketNumber}": {
      "delete": {
        "tags": [
          "ticket-controller"
        ],
        "operationId": "deleteTicket",
        "parameters": [
          {
            "name": "ticketNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AirportUpdateRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "AirportUpdateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "AirplaneUpdateRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "capacity": {
            "type": "integer",
            "format": "int64"
          },
          "airlineCompanyID": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AirlineCompanyGetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "AirplaneUpdateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "capacity": {
            "type": "integer",
            "format": "int64"
          },
          "airlineCompany": {
            "$ref": "#/components/schemas/AirlineCompanyGetResponse"
          }
        }
      },
      "AirlineCompanyUpdateRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "AirlineCompanyUpdateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "TicketSaveRequest": {
        "required": [
          "cardNumber"
        ],
        "type": "object",
        "properties": {
          "cardNumber": {
            "type": "string"
          },
          "flightId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AirplaneGetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "capacity": {
            "type": "integer",
            "format": "int64"
          },
          "airlineCompany": {
            "$ref": "#/components/schemas/AirlineCompanyGetResponse"
          }
        }
      },
      "AirportGetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          }
        }
      },
      "FlightGetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "flightNumber": {
            "type": "string"
          },
          "originAirport": {
            "$ref": "#/components/schemas/AirportGetResponse"
          },
          "destinationAirport": {
            "$ref": "#/components/schemas/AirportGetResponse"
          },
          "airplane": {
            "$ref": "#/components/schemas/AirplaneGetResponse"
          }
        }
      },
      "TicketSaveResponse": {
        "type": "object",
        "properties": {
          "ticketNumber": {
            "type": "string"
          },
          "cardNumber": {
            "type": "string"
          },
          "flight": {
            "$ref": "#/components/schemas/FlightGetResponse"
          }
        }
      },
      "FlightSaveRequest": {
        "required": [
          "flightNumber"
        ],
        "type": "object",
        "properties": {
          "flightNumber": {
            "type": "string"
          },
          "originAirportId": {
            "type": "integer",
            "format": "int64"
          },
          "destinationAirportId": {
            "type": "integer",
            "format": "int64"
          },
          "airplaneId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "FlightSaveResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "flightNumber": {
            "type": "string"
          },
          "originAirport": {
            "$ref": "#/components/schemas/AirportGetResponse"
          },
          "destinationAirport": {
            "$ref": "#/components/schemas/AirportGetResponse"
          },
          "airplane": {
            "$ref": "#/components/schemas/AirplaneGetResponse"
          }
        }
      },
      "AirportSaveRequest": {
        "required": [
          "city",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          }
        }
      },
      "AirportSaveResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          }
        }
      },
      "AirplaneSaveRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "capacity": {
            "type": "integer",
            "format": "int64"
          },
          "airlineCompanyID": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AirplaneSaveResponse": {
        "required": [
          "airlineCompany",
          "capacity",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "capacity": {
            "type": "integer",
            "format": "int64"
          },
          "airlineCompany": {
            "$ref": "#/components/schemas/AirlineCompanyGetResponse"
          }
        }
      },
      "AirlineCompanySaveRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "AirlineCompanySaveResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "TicketGetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "ticketNumber": {
            "type": "string"
          },
          "cardNumber": {
            "type": "string"
          },
          "flightId": {
            "$ref": "#/components/schemas/FlightGetResponse"
          }
        }
      }
    }
  }
}