# src/main/java/io/upschool/ticketSystem/dto/airlineCompany/AirlineCompanyUpdateRequest.java
package io.upschool.ticketSystem.dto.airlineCompany;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.Data;

@Data
public class AirlineCompanyUpdateRequest {

    private Long id;

    @NotBlank
    @NotEmpty
    @NotNull
    @Size(min = 2, max = 100,message = "Name alanı minimum 2 maksimum 100 karater olabilir.")
    private String name;
}